requirements.txt
alabaster==0.7.12
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.1.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
black==19.10b0
bootstrap4==0.1.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.2.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.8
django-allauth==0.42.0
django-anymail==7.2.1
django-bootstrap4==2.2.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==3.0.3
django-filter==2.3.0
django-guardian==2.3.0
django-model-utils==4.0.0
django-redis==4.12.1
django-stubs==1.5.0
django-storages[google]==1.9.1
django-tables2==2.3.1
django-widget-tweaks==1.4.8
docutils==0.16
factory-boy==2.12.0
Faker==4.1.2
filelock==3.0.12
flake8==3.8.3
flake8-isort==3.0.1
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-secret-manager==1.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
google-api-python-client==1.11.0
grpcio==1.31.0
gunicorn==20.0.4
identify==1.4.30
idna==2.10
imagesize==1.2.0
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
livereload==2.6.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
mypy==0.770
mypy-extensions==0.4.3
nodeenv==1.5.0
oauthlib==3.1.0
octicons==2.1.0
packaging==20.4
parso==0.7.1
pathspec==0.8.0
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
port-for==0.3.1
pre-commit==2.6.0
prompt-toolkit==3.0.7
protobuf==3.13.0
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.6.0
pylint-django==2.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.1.2
sphinx-autobuild==0.7.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
termcolor==1.1.0
testfixtures==6.14.1
text-unidecode==1.3
toml==0.10.1
tornado==6.0.4
traitlets==5.0.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.0.31
watchdog==0.10.3
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.1.0
wrapt==1.12.1
